backend/requirements.txt
PyMuPDF==1.23.0
sentence-transformers==2.7.0
chromadb==0.4.22
transformers==4.36.0
torch==2.1.0
pandas==2.1.4
numpy==1.24.3
fastapi==0.104.1
anthropic==0.39.0
pydantic==2.5.0
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
matplotlib==3.8.0